Explicitly require `oscar` in template Vagrantfile
Since the templated Vagrantfile has to reference the `Oscar` constant,
in order to read the Vagrantfile we have to require the oscar plugin.
Vagrant will normally load all plugins before reading the Vagrantfile,
but using the `vagrant plugin` subcommands will _not_ load plugins but
_will_ read the Vagrantfile. To avoid evaluating an uninitialized
constant we need to explicitly require this constant.

It's up for grabs whether using `Vagrant.require_plugin` is actually
better than `require`, but we'll have to see.